added functionality for a modal to appear on screen when a Pokémon button is clicked
Populated the "modal-container" element in HTML with Pokédex information.
Allowed for the modal to be closed in multiple ways.
Created a function outside of IIFE in order to only have showModalOnClick execute when an event listener click occurs, followed by a querySelector in the showModal function to remove the event listener once showModalOncClick has been executed.
Updated comments accordingly.